SwarmMC.jl

Documentation for SwarmMC.jl

Simple API list in order of use.

Setup of simulation

SwarmMC.PARAMS — Type

PARAMS(; kwds...)

The main parameters object. The parameters can be set as keywords or by mutating the structure afterwards. When finished setting the parameters, call Finalise(...) on the object which will set the structure type parameters and fill in the necessary interal-use-only fields.

Keywords that must be set (no defaults):

init_style - how particles are generated

t_grid - the time grid used for the simulation and for measurements

z_grid, r_grid - the grids used for measurements

gas_list - a list of GAS objects

ptype_list - a list of PARTTYPE objects

gen_cf - a function that returns the list of collision frequencies for each gas and particle type combination. Function signature (params, gas, ptype) -> ...

Keywords that have defaults:

eps_grid - the energy grid used for measurements

meas_bins - the bins for measurements

meas_quants - the quantities measured

steady_state_timefrac - defines the time at which steady-state measurements are taken from.

E_field - the electric field. Can be nothing, an XYZ vector, or a function of signature (pos,time) -> ....

B_field - the magnetic field. Can be nothing, an XYZ vector, or a function of signature (pos,time) -> ....

min_log2_weight - the point at which if a particle's weight is less than 2^min_log2_weight

weight_reduction - the percentage of the weight that is lost in a loss collision.

ionisation_as_weight - whether to either create a secondary particle in ionisation, or to double the weight.

generate_next_step_style - one of GNS_ENUM types for refreshing the particle list for each timestep.

only_regen_for_too_many - when true, don't ever generate extra particles when using GNS_REGEN_ALL.

max_substepsize - the maximum timestep. Use this when the time grid is too coarse to handle changing numbers of particles.

runaway_threshold - the multiplicative factor to judge when a runaway of particles is occuring, which will stop the simulation.

static_structure_factor - the static structure factor used for elastic collisions

save_name - the name for saving, see MakeSaveName

do_event_meas - if true, store information in event_counts

show_events - log whenever an event occurs.

handle_temperature_bias - set this to false to force sampling gas velocities directly (and incorrectly) from a Maxwellian distribtion.

user_callbacks - a list of additional callbacks to provide to the ODE solver. These callbacks will directly interface with DifferentialEquations.jl. For reference, the p in the integrator object will be of type INTEGRATOR.

SwarmMC.Finalise — Function

Finalise(params)

Turn a PARAMS object into the most narrow set of types and also fill in all internal-use parts of the object.

SwarmMC.PARTTYPE — Type

PARTTYPE(;name=:particle, mass=1mₑ, charge=1q)

The definition of the type of particle, giving a mass and charge.

SwarmMC.GAS — Type

GAS(; name::Symbol=:gas, m0, ρ, tmtr)

The definition of a gas. The temperature tmtr defaults to zero, which itself must be specified as nothing.

Both ρ and tmtr can accept a number or a function which itself takes one argument: pos.

SwarmMC.CreateCollFreq — Function

CreateCollFreq(params, gas, ptype, name, colltype::CF_STYLE_ENUM, gencf::GENCF_PARAMS ; threshold=0.0uE, ion_sharing_style=ISS_AFE(), ion_sharing_ratio=0.5, is_cross_section=true)

Generate a COLLFREQ object, should be used for creating the list to be given to gen_cf of the params object.

The collision type should be chosen from one of those in CF_STYLE_ENUM. The details of the form of the collision frequency should be provided with gencf. Options are:

GENCF_MANUAL - an arbitrary functional form of the cross section

GENCF_INTERP - interpolate a given list of energy and cross section

GENCFINTERPLOG - as with GENCFINTERP but interpolate on the log values

Some convenience structs:

GENCF_HARDSPHERE

GENCF_MAXWELL

GENCF_ATTACH - for the Ness-Robson attachment benchmark

By default, these are interpreted as a cross section. If a collision frequency is directly required, then set is_cross_section to false.

For inelastic cross sections, threshold is important. For ionisation cross sections, ion_sharing_style and ion_sharing_ratio is important.

DCS values can be provided using angle_dist_cum.

SwarmMC.CF_STYLE_ENUM — Type

CF_STYLE_ENUM

The type of collision frequency process.

These types are hopefully obvious:

CFS_ELASTIC

CFS_INELASTIC

CFS_LOSS

CFS_IONISATION

The types that might need some more explanation are:

CFS_NULL - for introducing manually configured null collisions.

CFSINELASTICMANUAL_DEGEN - an inelastic cross section with manually set ground/excitation populations, as opposed to using the gas temperature.

CFS_CHANGETYPE - changes the type of particle.

SwarmMC.ISS_ENUM — Type

ISS_ENUM

Options for ionisation energy sharing

ISS_FRACTION - specify an exact constant fraction

ISS_AFE - all fractions equiprobable

ISS_FUNC - the fraction is chosen at the time of collision.

SwarmMC.GNS_ENUM — Type

GNS_ENUM

The style for generating or removing particles for the next time bin.

GNS_NOTHING - don't ever generate or remove new particles

GNSUPDATELOG2FAC - as with GNS_NOTHING but update the log2fac of the particles.

GNS_DOUBLE - when there are too few particles, double those that are left.

GNSREGENALL - when the effective number of particles gets too high or low, regenerate by sampling the same weights of particles.

SwarmMC.MEAS_BIN — Type

MEAS_BIN([:cum][,:ss][,:t][,:r][,:z][,:eps][,:costh])

Define the times, positions or energies in which measurements are to be divided into. It also selects whether measurements should be cumulative or instantaneous.

The constructor takes symbols, e.g. MEAS_BIN(:cum,:t) represents cumulative measurements in time bins and MEAS_BIN(:ss,:eps,:r) are instantaneous measurements in energy and radial bins, only for the steady-state times, defined by steady_state_frac in PARAMS.

SwarmMC.MEAS_QUANT — Type

MEAS_QUANT(label::Symbol, unit, kind::Type=Float64)

References the quantity that is measured and will be placed into MEAS_BIN. A quantity defined this way should then extend InstMeas using label to implement the measure itself.

The storage of this quantity is defined by unit and kind. kind can be Float64 or XYZ.

SwarmMC.MakeSaveName — Function

MakeSaveName(prefix ; separate_dir=true, kwds...)

Make up a filename prefix to save as, which is: "prefix:kwd1=val1:kwd2=val2:..."

If separate_dir is true then prefix all of the filenames with "runs_prefix/"

Running of simulation

SwarmMC.Save — Function

Save(params, props ; mmap=false)

Find a new filename and save the params and props into it using JLD2.

Choose whether to save with IOStream or mmap from the keyword.

SwarmMC.LoopMaxTime — Function

LoopMaxTime(params, N ; walltime=:auto)

Runs simulations repeatedly until walltime (a Unitful quantity) is reached.

If walltime is :auto then try to lookup the environment variable SWARMMC_WALLTIME and interpret it as a time in minutes.

N and params are passed through to BunchedPropagate.

SwarmMC.BunchedPropagate — Function

BunchedPropagate(params, N ; return_part_list=false)

Run a simulation of N particles using params. Returns a PROPS_OUT object.

If you pass true to return_part_list then the initial and final particle lists will also be returned.

SwarmMC.PROPS_OUT — Type

PROPS_OUT

Structure to store measurements taken during a simulation. This struct supports convenient querying of different quantities for different bins.

The user should not need to manually create or populate this, but will need to query it to obtain results.

Loading/Combining results

SwarmMC.ReadAll — Function

ReadAll(prefix)

Load and combine all params and props from files belonging to a particular prefix, i.e. the part of the filenames before the "__*.jld2" part.

SwarmMC.GetAll — Function

GetAll(names... ; kwds...)

Read all sets of results in the current directory, extracting quantities given by names. Each name can be either:

a symbol, in which case a key from Quants is used.

a string, in which case a parameter from the filename prefix is used. See NameElements

a function, f(params,props,prefix,quants) which produces a value.

A name may also be a Pair, name => value, where value is one of the three choices above.

SwarmMC.CollectUp — Function

CollectUp()

Combine files from the same prefix into one file.

SwarmMC.NameElements — Function

NameElements(str) NameElements(str, index::String)

Split and identify the different parts of a prefix.

In the second form, return a specific part given by index.

SwarmMC.PrefixSets — Function

PrefixSets()

Return all the different unique sets of prefixes in the current directory.

Analysing results

SwarmMC.Quants — Function

Quants(params, props ; mbin, SI=true, n0_factors=false, include_errors=true)

Generate a dictionary of commonly used transport quantities from time averages in props. The type of measurement bin can be selected in mbin, although it must be either an instantaneous or cumulative time-only set of measurements.

n0_factors indicate that the appropriate density-scaling factors (e.g. n0*DT) will be given instead of the raw values.

SwarmMC.EpsFromVel — Function

EpsFromVel(vel, mass) EpsFromVel(vel, ::PARTTYPE) EpsFromVel(::PARTICLE)

Calculate the energy of a particle from the velocity and mass.

SwarmMC.ΔR — Function

ΔR(params)

Return the size of the radial measurement bins.

SwarmMC.RGrid — Function

RGrid(params)

Return the midpoint of the radial measurement bins.

SwarmMC.TGrid — Function

TGrid(params)

Return the midpoint of all time bins, which is the value relevant for cumulative measurements.

SwarmMC.TInstGrid — Function

TInstGrid(params)

Return the points at which instantaneous measurements correspond to.

SwarmMC.ΔT — Function

ΔT(params)

Return the size of the cumulative time measurement bins.

SwarmMC.EpsGrid — Function

EpsGrid(params)

Return the midpoint of the energy measurement bins.

SwarmMC.ΔEps — Function

ΔEps(params)

Return the size of the energy measurement bins.

SwarmMC.ZGrid — Function

ZGrid(params)

Return the midpoint of the spatial z measurement bins.

SwarmMC.ΔZ — Function

ΔZ(params)

Return the size of the spatial z measurement bins.

SwarmMC.CosthGrid — Function

CosthGrid(params)

Return the midpoint of the costheta measurement bins.

SwarmMC.ΔCosth — Function

ΔCosth(params)

Return the size of the costheta measurement bins.

Advanced

Definining new measurement types

SwarmMC.Log2Fac — Function

Log2Fac(::Val{sym}, log2fac)

What log2fac should be assigned to this quantity. Only relevant for measurements that are considering the weights.

SwarmMC.InstValue — Function

InstValue(::Val{sym}, weight, log2fac, pos, vel, eps)

The value of the quantity (sym) being measured, given the current particle properties are pos, vel, eps and the particle's weight is given by weight + 2^log2fac.

User callbacks

SwarmMC.INTEGRATOR — Type

INTEGRATOR

The parameter object passed to DifferentialEquations.jl to provide information about the simulation. The user should not need to care about this, unless they are trying to implement custom callbacks.

A custom callback will have access to this structure and hence the following properties:

params - the PARAMS used for the simulation

part - the current state of the particle

extra_part_list - any additional particles to pass back to the simulation (e.g. created during ionisation).

props_out - the measurements object

TODO

[ ] Custom measurements

[ ] CF + nonconservative measurements